+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This is used for recording Invader's changes. This changelog is based on
 - invader-compare: Verbose comparisons are now shown in an inverted tree, making
   it easier for collapsing in text editors
 - invader-compare: Verbose comparisons can now be threaded
+- invader-edit-qt: Changed default sizing of the main window to be larger
 
 ### Fixed
 - invader-build: Fixed a number of scripting issues on MCC. This notably fixes